+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added min width to compute plans table columns to fix display when there are no results (#229)
 -   Truncate TableDrawerSectionKeyEntry value when too long (#234)
 -   Set minWidth on a compute plan tag in order to avoid to many multilines (#235)
+-   Avoid NaN progression percentage (#238)
 
 ### Changed